Refactor service clients for test/main separation
Moved RestAssured-based service client implementations to src/test and replaced main module clients with stubs throwing UnsupportedOperationException. This enables integration tests to use real HTTP clients while keeping production code free of test dependencies. Also split UserRegistrationRequest into its own file and added rest-assured dependency to pom.xml.
