+++ integration-tests/src/main/java/com/taller/integration/client/ConsumerServiceClient.java
@@ -5,33 +5,18 @@ import io.restassured.response.Response;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
-import static io.restassured.RestAssured.given;
-
 @Component
 @RequiredArgsConstructor
 public class ConsumerServiceClient {
 
     private final IntegrationTestConfig config;
 
+    // Implementación stub para production/main module — las pruebas usan la clase en src/test
     public Response consumeApps(String nameForGreeting) {
-        return given()
-                .baseUri(config.getConsumerBaseUrl())
-                .pathParam("nameForGreeting", nameForGreeting)
-                .when()
-                .get("/consumeApps/{nameForGreeting}")
-                .then()
-                .extract()
-                .response();
+        throw new UnsupportedOperationException("Este cliente es solo un stub en el código main. Use la clase de pruebas en src/test para integración.");
     }
 
     public Response healthCheck() {
-        return given()
-                .baseUri(config.getConsumerBaseUrl())
-                .when()
-                .get("/api/health")
-                .then()
-                .extract()
-                .response();
+        throw new UnsupportedOperationException("Este cliente es solo un stub en el código main. Use la clase de pruebas en src/test para integración.");
     }
 }
-
