+++ integration-tests/src/main/java/com/taller/integration/client/OrchestratorServiceClient.java
@@ -2,12 +2,9 @@ package com.taller.integration.client;
 
 import com.taller.integration.config.IntegrationTestConfig;
 import com.taller.integration.dto.NotificationRequest;
-import io.restassured.http.ContentType;
 import io.restassured.response.Response;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-
-import static io.restassured.RestAssured.given;
 
 @Component
 @RequiredArgsConstructor
@@ -16,72 +13,26 @@ public class OrchestratorServiceClient {
     private final IntegrationTestConfig config;
 
     public Response createNotification(NotificationRequest request, String token) {
-        return given()
-                .baseUri(config.getOrchestratorBaseUrl())
-                .header("Authorization", "Bearer " + token)
-                .contentType(ContentType.JSON)
-                .body(request)
-                .when()
-                .post("/api/notifications")
-                .then()
-                .extract()
-                .response();
+        throw new UnsupportedOperationException("Stub en main; las pruebas usan la implementación en src/test que depende de RestAssured.");
     }
 
     public Response getNotifications(String token, int page, int size) {
-        return given()
-                .baseUri(config.getOrchestratorBaseUrl())
-                .header("Authorization", "Bearer " + token)
-                .queryParam("page", page)
-                .queryParam("size", size)
-                .when()
-                .get("/api/notifications")
-                .then()
-                .extract()
-                .response();
+        throw new UnsupportedOperationException("Stub en main; las pruebas usan la implementación en src/test que depende de RestAssured.");
     }
 
     public Response getNotificationById(String token, String notificationId) {
-        return given()
-                .baseUri(config.getOrchestratorBaseUrl())
-                .header("Authorization", "Bearer " + token)
-                .when()
-                .get("/api/notifications/" + notificationId)
-                .then()
-                .extract()
-                .response();
+        throw new UnsupportedOperationException("Stub en main; las pruebas usan la implementación en src/test que depende de RestAssured.");
     }
 
     public Response getChannels(String token) {
-        return given()
-                .baseUri(config.getOrchestratorBaseUrl())
-                .header("Authorization", "Bearer " + token)
-                .when()
-                .get("/api/notifications/channels")
-                .then()
-                .extract()
-                .response();
+        throw new UnsupportedOperationException("Stub en main; las pruebas usan la implementación en src/test que depende de RestAssured.");
     }
 
     public Response getTemplates(String token) {
-        return given()
-                .baseUri(config.getOrchestratorBaseUrl())
-                .header("Authorization", "Bearer " + token)
-                .when()
-                .get("/api/notifications/templates")
-                .then()
-                .extract()
-                .response();
+        throw new UnsupportedOperationException("Stub en main; las pruebas usan la implementación en src/test que depende de RestAssured.");
     }
 
     public Response healthCheck() {
-        return given()
-                .baseUri(config.getOrchestratorBaseUrl())
-                .when()
-                .get("/api/health")
-                .then()
-                .extract()
-                .response();
+        throw new UnsupportedOperationException("Stub en main; las pruebas usan la implementación en src/test que depende de RestAssured.");
     }
 }
-
