+++ integration-tests/src/main/java/com/taller/integration/dto/NotificationRequest.java
@@ -16,20 +16,3 @@ public class NotificationRequest {
     private Map<String, Object> parameters;
     private String message;
 }
-package com.taller.integration.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserRegistrationRequest {
-    private String username;
-    private String email;
-    private String password;
-    private String firstName;
-    private String lastName;
-}
-
