+++ integration-tests/src/main/java/com/taller/integration/dto/UserRegistrationRequest.java
@@ -1,0 +1,17 @@
+package com.taller.integration.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserRegistrationRequest {
+    private String username;
+    private String email;
+    private String password;
+    private String firstName;
+    private String lastName;
+}
+
